+++ examples/mcp-puppeteer/usages.ts
@@ -1,0 +1,37 @@
+import assert from "node:assert";
+import { AIAgent, ExecutionEngine, MCPAgent, OpenAIChatModel } from "@aigne/core";
+
+const { OPENAI_API_KEY } = process.env;
+assert(OPENAI_API_KEY, "Please set the OPENAI_API_KEY environment variable");
+
+const model = new OpenAIChatModel({
+  apiKey: OPENAI_API_KEY,
+});
+
+const puppeteerMCPAgent = await MCPAgent.from({
+  command: "npx",
+  args: ["-y", "@modelcontextprotocol/server-puppeteer"],
+});
+
+const engine = new ExecutionEngine({
+  model,
+  tools: [puppeteerMCPAgent],
+});
+
+const agent = AIAgent.from({
+  instructions: `\
+## Steps to extract content from a website
+1. navigate to the url
+2. evaluate document.body.innerText to get the content
+`,
+});
+
+const result = await engine.call(agent, "extract content from https://www.arcblock.io");
+
+console.log(result);
+// output:
+// {
+//   $message: "The content extracted from the website [ArcBlock](https://www.arcblock.io) is as follows:\n\n---\n\n**Redefining Software Architect and Ecosystems**\n\nA total solution for building decentralized applications ...",
+// }
+
+await engine.shutdown();