+++ examples/workflow-sequential/usages.ts
@@ -1,10 +1,10 @@
 import assert from "node:assert";
-import { AIAgent, ChatModelOpenAI, ExecutionEngine } from "@aigne/core-next";
+import { AIAgent, ExecutionEngine, OpenAIChatModel, sequential } from "@aigne/core";
 
 const { OPENAI_API_KEY } = process.env;
 assert(OPENAI_API_KEY, "Please set the OPENAI_API_KEY environment variable");
 
-const model = new ChatModelOpenAI({
+const model = new OpenAIChatModel({
   apiKey: OPENAI_API_KEY,
 });
 
@@ -52,12 +52,9 @@ Draft copy:
 
 const engine = new ExecutionEngine({ model });
 
-const result = await engine.run(
-  { product: "AIGNE is a No-code Generative AI Apps Engine" },
-  conceptExtractor,
-  writer,
-  formatProof,
-);
+const result = await engine.call(sequential(conceptExtractor, writer, formatProof), {
+  product: "AIGNE is a No-code Generative AI Apps Engine",
+});
 
 console.log(result);