+++ packages/core/src/utils/json-schema.ts
@@ -1,0 +1,39 @@
+import { isObject } from "lodash-es";
+import type { ZodType } from "zod";
+import { zodToJsonSchema } from "zod-to-json-schema";
+import type { Message } from "../agents/agent.js";
+import { logger } from "./logger.js";
+
+export function outputSchemaToResponseFormatSchema(
+  agentOutput: ZodType<Message>,
+): Record<string, unknown> {
+  return setAdditionPropertiesDeep(zodToJsonSchema(agentOutput), false);
+}
+
+function setAdditionPropertiesDeep<T>(schema: T, additionalProperties: boolean): T {
+  if (Array.isArray(schema)) {
+    return schema.map((s) => setAdditionPropertiesDeep(s, additionalProperties)) as T;
+  }
+  if (isObject(schema)) {
+    return Object.entries(schema).reduce(
+      (acc, [key, value]) => {
+        acc[key] = setAdditionPropertiesDeep(value, additionalProperties);
+        if (acc.type === "object") {
+          acc.additionalProperties = additionalProperties;
+        }
+        return acc;
+      },
+      {} as Record<string, unknown>,
+    ) as T;
+  }
+  return schema;
+}
+
+export function parseJSON(json: string) {
+  try {
+    return JSON.parse(json);
+  } catch (error) {
+    logger.debug("Failed to parse JSON", { json, error });
+    throw new Error(`Failed to parse JSON ${error.message}`);
+  }
+}