+++ packages/core/test/agents/mcp-agent.test.ts
@@ -1,6 +1,6 @@
 import { expect, test } from "bun:test";
 import { join } from "node:path";
-import { MCPAgent } from "@aigne/core-next";
+import { MCPAgent } from "@aigne/core";
 
 test("MCPAgent should correctly call tool, get prompt and read resource", async () => {
   const mcp = await MCPAgent.from({
@@ -9,6 +9,9 @@ test("MCPAgent should correctly call tool, get prompt and read resource", async 
   });
 
   try {
+    expect(mcp.isCallable).toBe(false);
+    expect(mcp.call({})).rejects.toThrowError();
+
     expect(mcp.tools.map((i) => i.name)).toEqual(["echo"]);
     expect(await mcp.tools.echo?.call({ message: "AIGNE" })).toEqual(
       expect.objectContaining({