+++ CHANGELOG.md
@@ -1,6 +1,34 @@
+## [1.2.0](https://github.com/AIGNE-io/aigne-framework/compare/aigne-framework-v1.1.0...aigne-framework-v1.2.0) (2025-03-18)
+
+- chore: release v1.2.0
+
+
+## [1.3.1](https://github.com/AIGNE-io/aigne-framework/compare/aigne-framework-v1.3.0...aigne-framework-v1.3.1) (2025-03-25)
+
+
+### Bug Fixes
+
+* **core:** use system message as user message for claude model if needed ([#32](https://github.com/AIGNE-io/aigne-framework/issues/32)) ([316a6d5](https://github.com/AIGNE-io/aigne-framework/commit/316a6d51f885cceee4020c24695f6588f6b7425f))
+* simplify mcp examples ([2b0b50e](https://github.com/AIGNE-io/aigne-framework/commit/2b0b50ef278edb938d26e5b31f0be73eb51e9204))
+
+## [1.3.0](https://github.com/AIGNE-io/aigne-framework/compare/aigne-framework-v1.2.0...aigne-framework-v1.3.0) (2025-03-24)
+
+
+### Features
+
+* **core:** add ChatModelClaude to use models of anthropic ([#30](https://github.com/AIGNE-io/aigne-framework/issues/30)) ([0a62a64](https://github.com/AIGNE-io/aigne-framework/commit/0a62a6499e3da723a4646e67952051708ce7de6a))
+* **core:** add support for subscribing topics for agent memory ([#28](https://github.com/AIGNE-io/aigne-framework/issues/28)) ([eeecc67](https://github.com/AIGNE-io/aigne-framework/commit/eeecc67049a60ebcc4cdba0fbcd987b3d81f4af6))
+
+
+### Bug Fixes
+
+* **core:** enforce stricter input/output type checks ([#26](https://github.com/AIGNE-io/aigne-framework/issues/26)) ([ef8cf53](https://github.com/AIGNE-io/aigne-framework/commit/ef8cf53586aff08a809909c56ab2a20f215fa129))
+* rename @aigne/core-next to @aigne/core ([3a81009](https://github.com/AIGNE-io/aigne-framework/commit/3a8100962c81813217b687ae28e8de604419c622))
+
 ## 1.1.0-beta.17 (2025-3-18)
 
 - chore: add support for esm module
+
 
 ## 1.1.0-beta.16 (2025-3-18)
 
@@ -44,7 +72,7 @@
 
 ## 1.1.0-beta.7 (2025-3-13)
 
-- chore: rename @aigne/core to @aigne/core-next
+- chore: rename @aigne/core to @aigne/core
 
 ## 1.1.0-beta.6 (2025-3-13)